Move notes.json and services.js to data directory
