+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
 import GlobalStyle from './GlobalStyle'
 import NotesList from './components/NotesList'
-import { getNotes } from './services'
+import { getNotes } from './data/services'
 
 function App() {
   const notes = getNotes()
